Develop 2 (#3)
* prueba 1, solo agregue // para probar

* le quite lo de prueba 1, para dejarlo a como estaba antes de tocar el
codigo

* cree la clase Conexion con recibos para que todo esta bien separado, y
asi dejar la conexion de recibos con BD aislada

* cree las clases de conexionConTablaPagos y ConexionConRecibos para
separar todo muy bien, y elimine los metodos existentes en conexionSQL
para q estuvieran donde debian estar

* se crearon las clases de LogicaDeNegocioDePersona y
LogicaDeNegocioDeOperaciones para crear los metodos relacionados con las
operacion fuera de la BD
