+++ CemprodecaDataManager/src/LogicaDeNegocio/LogicaDeNegocioDeOperaciones.java
@@ -1,0 +1,5 @@
+package LogicaDeNegocio;
+
+public class LogicaDeNegocioDeOperaciones {
+
+}
