+++ CemprodecaDataManager/src/LogicaDeNegocio/LogicaDeNegocioDePersona.java
@@ -1,0 +1,26 @@
+package LogicaDeNegocio;
+
+import estructuraDeDatos.Persona;
+
+public class LogicaDeNegocioDePersona {
+
+	Persona LaPersona;
+	
+	public Persona BuscarPersonaPorCedula(String laCedula)
+	{
+	Persona laPersonaBuscada;
+	
+	
+	
+	
+	return laPersonaBuscada;	
+		
+	}
+	
+	
+	
+	
+	
+	
+
+}
