+++ CemprodecaDataManager/src/conexionesBD/ConexionConRecibos.java
@@ -1,0 +1,22 @@
+package conexionesBD;
+
+import java.util.ArrayList;
+
+import estructuraDeDatos.Recibos;
+
+public class ConexionConRecibos {
+
+	
+	public ArrayList<Recibos> DescargaRecibos(){
+		System.out.println("NO ESTA DEFINIDO EL CONTENIDO DEL METODO");
+	return null;/*retorna los recibos viejos de la db*/
+	}/*fin de Recibos entrantes*/
+	
+
+	
+	public void CargarRecibosAlaDB(){
+		
+	}/*fin de carga de recibos*/
+	
+	
+}
