+++ CemprodecaDataManager/src/conexionesBD/ConexionConTablaPagos.java
@@ -1,0 +1,20 @@
+package conexionesBD;
+
+import java.util.ArrayList;
+
+import estructuraDeDatos.TablaDePagos;
+
+public class ConexionConTablaPagos {
+
+	public ArrayList<TablaDePagos> DescargaTablaDePagos(){
+		System.out.println("NO ESTA DEFINIDO EL CONTENIDO DEL METODO");
+		return null;/*retorna un arraylist sacado de la db con las filas de la tabla de pagos*/
+		}/*fin de TablaDePagos entrante*/
+
+	
+	public void CargarTablaDePagosAlaDB(){
+		
+	}/*fin de carga de tabla de pagos*/
+	
+	
+}
